Add files via upload
This is important in any integration to convert the payload as per the target system requirements.

To convert the file format we have multiple ways one of them is to use Convertor function where we have different pre-defined scenarios.

We are using JSON to XML.

For sending the payload we are using HTTP after that add Convertor with JSON to XML.

Click on blank space In Integration flow -> Runtime Configuration -> Namespace Mapping and put Namespace values.

Now how to create Namespace follow the below steps 

1. Namespace Format:
A namespace is typically defined using a URI (Uniform Resource Identifier), which is usually in the form of a URL, though it doesn't have to be an active, reachable address. The format looks like this: xmlns:prefix="http://www.example.com/namespace"

xmlns is the standard XML namespace declaration keyword.
prefix is an optional alias you can assign to the namespace, like ns or order, which can be used to refer to the namespace in the XML document.
"http://www.example.com/namespace" is the URI (a unique identifier) for the namespace.

2. Why Use a Namespace:
Namespaces are necessary to:

Differentiate elements with the same name but different meanings.
Avoid conflicts when mixing elements from different XML vocabularies or data models.
For example, if two different schemas both define an element <Order>, the namespace ensures the system knows which one to interpret correctly.

3. Creating a Namespace:
Here’s how you can create a namespace:

The namespace URI should be unique to your application or organization.
You can use your organization's domain name as the URI, though it doesn’t need to be an actual URL that leads to a webpage.
Use the namespace in your XML document by declaring it in the root element or relevant elements.
 
JSON To XML Converter:

Uncheck the Namespace Mapping as we are not using it in this example then inside Name put root element and select the Namespace mapping which we defined in previous step.

FInish.